Pulled duplicated functions into common base class
